Moved motor firmware-specific headers to the include section
