+++ stm32-modules/include/thermocycler-refresh/firmware/motor_hardware.h
